自测通过(snapshot暂时无法复现获取不到的情况)

+++ dev-sdk/src/com/analysys/track/model/PolicyInfo.java
@@ -79,7 +79,7 @@ public class PolicyInfo implements Serializable {
         return PolicyVer;
     }
 
-    public long getServerDelay() {
+    public int getServerDelay() {
         return ServerDelay;
     }
